edge_server/app/requirements.txt
alembic==1.0.7
backports.functools-lru-cache==1.5
beautifulsoup4==4.5.3
blinker==1.3
chardet==2.3.0
Click==7.0
colorama==0.3.7
cryptography==1.7.1
cycler==0.10.0
enum34==1.1.6
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
html5lib==0.999999999
idna==2.2
imutils==0.5.2
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.10
keyring==10.1
keyrings.alt==1.3
kiwisolver==1.0.1
lxml==3.7.1
Mako==1.0.7
MarkupSafe==1.1.0
matplotlib==2.2.3
mpmath==1.1.0
numpy==1.15.4
opencv-python==3.4.5.20
Pillow==5.4.1
psutil==5.0.1
pyasn1==0.1.9
pycrypto==2.6.1
pyinotify==0.9.6
pyOpenSSL==16.2.0
pyparsing==2.3.1
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
pyxdg==0.25
PyYAML==3.13
requests==2.12.4
RPi.GPIO==0.6.5
scour==0.32
SecretStorage==2.3.1
simplejson==3.10.0
six==1.12.0
SQLAlchemy==1.2.17
style==1.1.0
subprocess32==3.5.3
sympy==1.3
urllib3==1.19.1
virtualenv==15.1.0
webencodings==0.5
Werkzeug==0.14.1
